requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
umap-learn>=0.5.0
zarr>=2.11.0
jupyter>=1.0.0
ipykernel>=6.0.0
openpyxl>=3.0.0
